+++ lib/src/http/base_middleware.dart
@@ -1,9 +1,0 @@
-library angel_framework.http.base_middleware;
-
-import 'dart:async';
-import 'request_context.dart';
-import 'response_context.dart';
-
-abstract class AngelMiddleware {
-  Future<bool> call(RequestContext req, ResponseContext res);
-}