+++ lib/src/http/base_plugin.dart
@@ -1,8 +1,0 @@
-library angel_framework.http.base_plugin;
-
-import 'dart:async';
-import 'server.dart';
-
-abstract class AngelPlugin {
-  Future call(Angel app);
-}