+++ lib/src/http/http.dart
@@ -10,8 +10,6 @@ export 'package:angel_route/angel_route.dart';
 export 'package:body_parser/body_parser.dart' show FileUploadInfo;
 export 'angel_base.dart';
 export 'anonymous_service.dart';
-export 'base_middleware.dart';
-export 'base_plugin.dart';
 export 'controller.dart';
 export 'hooked_service.dart';
 export 'map_service.dart';
@@ -24,12 +22,12 @@ export 'service.dart';
 export 'typed_service.dart';
 
 /// Boots a shared server instance. Use this if launching multiple isolates
-Future<HttpServer> startShared(InternetAddress address, int port) => HttpServer
+Future<HttpServer> startShared(address, int port) => HttpServer
     .bind(address ?? InternetAddress.LOOPBACK_IP_V4, port ?? 0, shared: true);
 
 /// Boots a secure shared server instance. Use this if launching multiple isolates
 ServerGenerator startSharedSecure(SecurityContext securityContext) {
-  return (InternetAddress address, int port) => HttpServer.bindSecure(
+  return (address, int port) => HttpServer.bindSecure(
       address ?? InternetAddress.LOOPBACK_IP_V4, port ?? 0, securityContext,
       shared: true);
 }