+++ lib/src/http/request_context.dart
@@ -5,14 +5,12 @@ import 'dart:io';
 import 'dart:mirrors';
 import 'package:body_parser/body_parser.dart';
 import 'package:charcode/charcode.dart';
-import '../fast_name_from_symbol.dart';
 import 'response_context.dart';
 import 'routable.dart';
 import 'server.dart' show Angel;
 part 'injection.dart';
 
 /// A convenience wrapper around an incoming HTTP request.
-@proxy
 class RequestContext {
   String _acceptHeaderCache;
   bool _acceptsAllCache;
@@ -277,23 +275,5 @@ class RequestContext {
       _provisionalQuery = null;
     return _body = await parseBody(io,
         storeOriginalBuffer: app.storeOriginalBuffer == true);
-  }
-
-  operator [](key) => properties[key];
-  operator []=(key, value) => properties[key] = value;
-
-  noSuchMethod(Invocation invocation) {
-    if (invocation.memberName != null) {
-      String name = fastNameFromSymbol(invocation.memberName);
-
-      if (invocation.isMethod) {
-        return Function.apply(properties[name], invocation.positionalArguments,
-            invocation.namedArguments);
-      } else if (invocation.isGetter) {
-        return properties[name];
-      }
-    }
-
-    return super.noSuchMethod(invocation);
   }
 }